+++ lib/pah/files/spec/acceptance/dummy_spec.rb
@@ -1,8 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe 'Homepage', type: :request do
-  xit do
-    visit root_path
-    page.should have_content('Copyright &copy; 2012 App')
-  end
-end